2 часть дипломного проекта

+++ stats/common-dto/src/main/java/ru/practicum/statistics_service/dto/ViewStatsDto.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class ViewStatsDto {
     private final String app;
     private final String uri;
-    private final Long hits;
+    private final Integer hits;
 }
